fix #5927: firewall: alias: lowercase names of aliases
pve-firewall lowercases the names of aliases when reading from the
configuration as well as when comparing source / destination entries
with the entries in the parsed aliases. In order to stay
backwards-compatible we also need to lowercase any parsed alias name.
I decided to this in the constructor and switch all call sites to the
new constructor, so there's only one place where we have to handle
lowercasing the string.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>