frr: add ospf types
Add OSPF-specific FRR types. This also reuses the types from
proxmox-network-types.

The NetworkType FRR option is implemented here, but not exposed to the
interface, as we want to keep it simple for the users. If they do not
set an IP, then the interface is considered to be unnumbered and uses
the Point-to-Point network type.

Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-12-g.goller@proxmox.com