frr: add route-map types
Only a very limited subset of the FRR route-maps is implemented here
intially, only what is currently needed for the fabrics feature. Once
standalone route-maps will make it into PVE, we will build on the
structs defined here and add possibly the full featureset. The main
use-case for routemaps in the fabrics is currently to overwrite the
source address in the kernel routing table, so that packets sent via
the fabric contain the router IP rather than the IP of the link (in
the case of point-to-point connections).

Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-13-g.goller@proxmox.com