config: sdn: fabrics: add section types
The fabrics configuration consists of two different entity types:
fabrics and nodes. This commit adds support for the fabrics inside
this section configuration.

There are properties required for all fabrics, regardless of type,
which are the ID of the fabric as well as the IP prefix of the fabric.
The FabricSection contains those shared properties, in order to avoid
having to duplicate them and any methods requiring them for every
protocol. This means that every concrete fabric section type is just
an instance of the generic type FabricSection<T>.

The type parameter is used to define properties, that are specific to
a protocol (e.g. area in OSPF). We also provide a generic
implementation for e.g. ApiType without having to avoid having to
repeat the API definitions and Updater types for the common properties
multiple times.

We also create the necessary types for updating a FabricSection<T> via
FabricSectionUpdater and FabricDeletableProperties, and provide
generic implementations for ApiType / Updater on FabricSection /
FabricSectionUpdater where possible.

This design allows to add new protocols simply by defining their
protocol-specific properties in a struct and then using it with the
generic FabricSection type.

The ID of a fabric is at most 8 alphanumeric characters long and can
additionally include hyphens. It cannot start or end with a hyphen.
This is because the name of the fabric, will be used as name for
network interfaces, which can be at most 15 characters long on Linux.
This leaves us enough space for generating interfaces with
pre/suffixes. This is analogous to the IDs of many other SDN entities,
e.g. VNet where the same restrictions apply.

The structs here will later be used in parsing the whole configuration
file, located in '/etc/pve/sdn/fabrics.cfg'.

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-16-g.goller@proxmox.com