config: sdn: fabrics: add node section types
NodeSection functions identically to the FabricSection type. It
contains all the common properties that nodes from all protocols have.
Protocol-specific properties can be defined via the type parameter of
NodeSection. It also provides generic implementations for ApiType, so
if the type parameter implements ApiType, then NodeSection<T> also
implements ApiType.

Together, FabricSection and NodeSection represent the two different
types of entities in the fabric section configuration, fabrics and
nodes.

IP addresses are optional because this enables nodes to be part of a
fabric without advertising an IP themselves. This enables nodes to
import routes from the fabric without announcing a route to
themselves. Also, since there can be either IPv4 or IPv6 (or both)
set, they have to be optional anyway.

The ID of a node is defined as the hostname of a node in the fabric,
but since nodes can be part of multiple fabrics their section config
entry can only be uniquely identified by a combination of the ID of
the fabric they belong to and the ID of the node. For this reason, the
ID of a node in the section config consists of the ID of the fabric as
well as the ID of the node, separated by an underscore. We provide a
helper struct for parsing the section ID into its two separate
components, so we can easily parse the ID on deserializing the section
config and easily serialize it back into its composite form when
serializing.

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-17-g.goller@proxmox.com