config: sdn: fabrics: add openfabric properties
This commit adds the protocol-specific properties, that are required
for an Openfabric fabric. They correspond to the respective properties
in the FRR fabricd configuration. For more information, see the FRR
documentation [1].

While the hello interval and csnp interval could be set on a
per-interface basis, it is recommended to keep them the same across
the whole fabric. This is why they can only be configured globally and
the value will be applied to all interfaces that are part of the
fabric. We expose the hello multiplier in the interface properties, so
users can define a longer hello interval on a per-interface basis by
supplying the hello multiplier parameter. The upside of this is, that
everything scales from a single value and users can just edit the
hello interval in the fabric and everything else will adjust based on
that setting.

We also introduce two new general enums: Fabric and Node. They contain
the concrete FabricSection and NodeSection types that are used for
each protocol and add the Openfabric sections to it. We provide
dedicated updater structs for the enum as well, that fall back to the
concrete updater structs of the respective variant.

New protocols can simply be added by adding another variant
to the Fabric and Node enums.

[1] https://docs.frrouting.org/en/latest/fabricd.html

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-19-g.goller@proxmox.com