config: sdn: fabrics: add ospf properties
Define the protocol-specific properties for OSPF and add the concrete
section types to the Fabric and Node enum. Currently only area is
included, which is also the only property that is required by FRR. We
wanted to start with a minimal set of properties and add any options
later on, depending on feedback from users.

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-20-g.goller@proxmox.com