config: sdn: fabrics: add api types
Add an api submodule to the section config module, that provides the
types that are intended to be returned and accepted by the Perl API in
Proxmox VE. This allows us to decouple the format returned in the API
from the configuration format.

This is particularly relevant in the case of the NodeSection type.
While the section config stores the composite ID of the node as the ID
of the section in the section config (and therefore as a single string
/ property), we want to be able to return them as independent fields
from the API, to avoid having to parse the ID everywhere else we want
to use it. Thanks to the generic NodeSection type we only have to
define the conversion from / to the API type once, while the
protocol-specific types can stay the same.

For the fabrics, we simply re-use the section_config types for now,
but by re-exporting them as type alias we are more flexible in
possibly changing the API types or the underlying section config types
later on.

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-21-g.goller@proxmox.com