config: sdn: fabrics: add section config
The Section type represents the configuration as stored in
'/etc/pve/sdn/fabrics.cfg'. It can contain a mix of fabric and node
sections, which are in turn unique to each protocol. We utilize the
generic FabricSection and NodeSection types to define every possible
section type in the section config. We also provide a helper for
sorting the sections in the configuration file into Fabrics and Nodes,
as well as conversion from / to the Fabric and Node types.

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-22-g.goller@proxmox.com