config: sdn: fabrics: add fabric config
The FabricConfig transforms the flat section configuration into its
hierarchical representation, which makes querying and validating the
fabrics configuration more ergonomic.

It provides the CRUD methods for safely manipulating the fabric
configuration, while checking for possible errors. It is intended to
be the interface for external users to use the fabric configuration.

By encapsulating the configuration into this struct, we can always
assure that invariants are upheld and many of them can actually be
checked at compile time (e.g. proper combination of FabricSection and
NodeSection in the hierarchy via the Entry struct).

It uses the Fabric and Node enums foremost in its public API, so
adding new protocols does not change the public API. This enables us
to write generic API methods that do not need to be updated when
adding new protocols.

If so desired, users can still access the protocol-specific properties
by matching on the FabricEntry enum.

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-23-g.goller@proxmox.com