ve-config: add optional tag property to vnet
When parsing the vnet.cfg config file we also want to get the tag
(vni or vlan). We need this so that we can lookup the vni of the passed
vnet on the status api call. We need the vni so that we can filter the
frr output.

In the future we would probably want to do this better with an enum per
vnet type and parse all possible options.

Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Tested-by: Hannes Dürr <h.duerr@proxmox.com>
Link: https://lore.proxmox.com/20251107143201.689035-7-s.hanreich@proxmox.com